unite ThemeComponent and ThemeContext
