--- src/contexts/ThemeContext.ts
@@ -1,6 +1,0 @@
-import { createContext } from 'react';
-
-export const ThemeContext = createContext({
-    theme: 'light' as 'light' | 'dark',
-    toggleTheme: () => {},
-});
